+++ CHANGELOG.md
@@ -1,13 +1,20 @@
 
 ## Todo/Issues
-- [ ] Currently modified buffers show/don't show tags incorrectly
-- [ ] Display in a way you can select the file and THEN pick the function/method under the file.
 - [ ] Custom "nearest tag function" for python to handle nested functions/methods.
 - [ ] Add gif to show usage.
 - [ ] Python 3 Compatible
 - [ ] Windows Compatible
 - [ ] Syntax/Colors for Browser 
+- [ ] Add by-class sorting into Expand Menu
+- [ ] Add "pick function by number"
 
+### 2013-10-20
+- [x] Expand Menu Display (by file, and function)
+- [x] Refactored
+
+### 2013-10-17
+- [x] Currently modified buffers show/don't show tags incorrectly
+- [x] Changed Folder Structure
 
 ### 2013-10-16
 - [x] Fix "wrongly logged tag due to file being edited" bug.